+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Updated loot sheet creation macro (!238 by mkahvi)
 - Added damage immunity and vulnerability to the Apply Damage dialog (!236 by mkahvi) (#995)
 - Add modifiers in breakdown to ability test rolls (!234 by mkahvi)
+- Add toggleable caster level checks to spells and their chat cards (!230 by mkahvi) (#333)
 
 ### Bug Fixes